services/nlp/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.5.2
hdbscan==0.8.38.post2
sentence-transformers==2.7.0
transformers==4.41.2
torch==2.2.2
keybert==0.8.5
yake==0.4.8
vaderSentiment==3.3.2
